Check if the user is the owner of the deal before removing it

+++ src/main/java/com/halildurmus/hotdeals/deal/DealServiceImpl.java
@@ -5,6 +5,8 @@ import static org.springframework.data.mongodb.core.query.Query.query;
 
 import com.halildurmus.hotdeals.deal.es.EsDeal;
 import com.halildurmus.hotdeals.deal.es.EsDealRepository;
+import com.halildurmus.hotdeals.security.SecurityService;
+import com.halildurmus.hotdeals.user.User;
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
 import org.bson.types.ObjectId;
@@ -29,6 +31,9 @@ public class DealServiceImpl implements DealService {
   @Autowired
   private MongoTemplate mongoTemplate;
 
+  @Autowired
+  private SecurityService securityService;
+
   @Transactional
   @Override
   public Deal saveOrUpdateDeal(Deal deal) {
@@ -40,18 +45,25 @@ public class DealServiceImpl implements DealService {
 
   @Transactional
   @Override
-  public void removeDeal(String id) {
+  public void removeDeal(String id) throws Exception {
+    final Deal deal = repository.findById(id)
+        .orElseThrow(() -> new Exception("Deal could not be found!"));
+    final User user = securityService.getUser();
+    if (!user.getId().equals(deal.getPostedBy().toString())) {
+      throw new Exception("You can only remove your own deal!");
+    }
+
     repository.deleteById(id);
     esDealRepository.deleteById(id);
   }
 
   @Override
   public Deal incrementViewsCounter(String dealId) throws Exception {
-    Query query = query(where("_id").is(dealId));
-    Update update = new Update().inc("views", 1);
-    FindAndModifyOptions options = FindAndModifyOptions.options().returnNew(true);
+    final Query query = query(where("_id").is(dealId));
+    final Update update = new Update().inc("views", 1);
+    final FindAndModifyOptions options = FindAndModifyOptions.options().returnNew(true);
 
-    Deal deal = mongoTemplate.findAndModify(query, update, options, Deal.class);
+    final Deal deal = mongoTemplate.findAndModify(query, update, options, Deal.class);
     if (deal == null) {
       throw new Exception("Deal could not be found!");
     }
@@ -60,7 +72,9 @@ public class DealServiceImpl implements DealService {
   }
 
   @Override
-  public Deal vote(String dealId, ObjectId userId, String voteType) throws Exception {
+  public Deal vote(String dealId, String voteType) throws Exception {
+    final User user = securityService.getUser();
+    final ObjectId userId = new ObjectId(user.getId());
     final Deal deal = repository.findById(dealId)
         .orElseThrow(() -> new Exception("Deal could not be found!"));
     final List<ObjectId> upvoters = deal.getUpvoters();
